Release of v0.0.2 (#25)
Development Release v0.0.2 patches up this libraries Javadocs. The official documentation can be viewed at: jbrew.org

===============================================================

* Updates for JavaDoc documentation.

* added example standard nrt under new package - standard.

* updated Task.java for next release.

* updated RetrievableTask.java for javadocs

* updates for release v0.0.2

* codecov update for release v0.0.2

* Update README.md
