--- src/main/java/org/jbrew/concurrent/exception.java
@@ -1,8 +1,0 @@
-package org.jbrew.concurrent;
-
-/**
- * An exception is thrown.
- * @author nealk
- *
- */
-public @interface exception {}
