--- src/main/java/org/jbrew/concurrent/param.java
@@ -1,7 +1,0 @@
-package org.jbrew.concurrent;
-
-/**
- * Input parameter.
- *
- */
-public @interface param {}
