+++ src/main/java/org/jbrew/concurrent/standard/StandardNonRetrievableTask.java
@@ -1,0 +1,17 @@
+package org.jbrew.concurrent.standard;
+
+import org.jbrew.concurrent.NonRetrievableTask;
+
+/**
+ * This is a standard implementation of the NonRetrievableTask.
+ * @author Neal Kumar
+ *
+ */
+public class StandardNonRetrievableTask extends NonRetrievableTask{
+
+	@Override
+	protected void execute() {
+		
+	}
+
+}
