+++ src/test/java/org/jbrew/concurrent_tests/nonretrievable/NonRetrievableTaskUnitTest.java
@@ -3,7 +3,7 @@ package org.jbrew.concurrent_tests.nonretrievable;
 import static org.junit.Assert.assertNotNull;
 
 import org.jbrew.Testing;
-import org.jbrew.concurrent.nonretrievable.NonRetrievableTask;
+import org.jbrew.concurrent.NonRetrievableTask;
 import org.junit.Test;
 
 @Testing
@@ -51,6 +51,8 @@ public class NonRetrievableTaskUnitTest {
 	
 	private class NR extends NonRetrievableTask{
 		@Override
-		protected void execute() {}
+		protected void execute() {
+			System.out.println("NonRetrievableTask works correctly.");
+		}
 	}
 }
